+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Introduced csv2db icon
 
+### Fixed
+- Pass correct exit return codes (bug #18)
+
 ## [1.3.1] - 2019-07-28
 
 ### Fixed